+++ lib/src/model/model.dart
@@ -14,3 +14,5 @@ export 'package:onfido_sdk/src/model/proof_of_address_result.dart';
 export 'package:onfido_sdk/src/model/enterprise_features.dart';
 export 'package:onfido_sdk/src/model/cobranding_logo.dart';
 export 'package:onfido_sdk/src/model/ios_appearance.dart';
+export 'package:onfido_sdk/src/model/onfido_theme.dart';
+export 'package:onfido_sdk/src/model/background_color.dart';