+++ lib/src/onfido.dart
@@ -9,19 +9,22 @@ class Onfido {
   /// - The [mediaCallback] allows you to receive captured results and handle them in your application.
   /// - The [enterpriseFeatures] is a set of features for advanced customisation, get in touch to enable this for your account.
   /// - The [disableNFC] is a flag used to disable NFC (now enabled by default)
+  /// - The [onfidoTheme] is an enum used to determine which theme the UI will be rendered with (DARK, LIGHT, AUTOMATIC)
   Onfido(
       {required String sdkToken,
       String? iosLocalizationFileName,
       IOSAppearance? iosAppearance,
       EnterpriseFeatures? enterpriseFeatures,
       bool? disableNFC,
-      OnfidoMediaCallback? mediaCallback})
+      OnfidoMediaCallback? mediaCallback,
+      OnfidoTheme? onfidoTheme})
       : _sdkToken = sdkToken,
         _iOSLocalizationFileName = iosLocalizationFileName,
         _enterpriseFeatures = enterpriseFeatures,
         _iosAppearance = iosAppearance,
         _disableNFC = disableNFC,
-        _mediaCallback = mediaCallback;
+        _mediaCallback = mediaCallback,
+        _onfidoTheme = onfidoTheme;
 
   final String _sdkToken;
   final String? _iOSLocalizationFileName;
@@ -29,6 +32,7 @@ class Onfido {
   final IOSAppearance? _iosAppearance;
   final bool? _disableNFC;
   final OnfidoMediaCallback? _mediaCallback;
+  final OnfidoTheme? _onfidoTheme;
 
   /// Start Onfido SDK using [FlowSteps].
   /// - The [flowSteps] allows you to configure which screens are going to be displayed in the flow. For more information see the [Start the Flow](https://github.com/onfido/flutter-sdk/#start-the-flow) section.
@@ -40,7 +44,8 @@ class Onfido {
         iosLocalizationFileName: _iOSLocalizationFileName,
         enterpriseFeatures: _enterpriseFeatures,
         disableNFC: _disableNFC,
-        mediaCallback: _mediaCallback);
+        mediaCallback: _mediaCallback,
+        onfidoTheme: _onfidoTheme);
   }
 
   /// Start Onfido SDK using [Workflow].
@@ -51,6 +56,7 @@ class Onfido {
         iosAppearance: _iosAppearance,
         mediaCallback: _mediaCallback,
         iosLocalizationFileName: _iOSLocalizationFileName,
-        enterpriseFeatures: _enterpriseFeatures);
+        enterpriseFeatures: _enterpriseFeatures,
+        onfidoTheme: _onfidoTheme);
   }
 }