+++ lib/src/serializer/start_studio_serializer.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:onfido_sdk/onfido_sdk.dart';
 
 /// convert dart code to dictionary to be passed to the paltform
@@ -8,14 +9,16 @@ class StartStudioSerializer {
       IOSAppearance? iosAppearance,
       String? iosLocalizationFileName,
       bool? shouldUseMediaCallback,
-      EnterpriseFeatures? enterpriseFeatures}) {
+      EnterpriseFeatures? enterpriseFeatures,
+      OnfidoTheme? onfidoTheme}) {
     return {
       'sdkToken': sdkToken,
       'workflowRunId': workflowRunId,
       'iosAppearance': iosAppearance?.toJson(),
       'iosLocalizationFileName': iosLocalizationFileName,
       'shouldUseMediaCallback': shouldUseMediaCallback,
-      'enterpriseFeatures': enterpriseFeatures?.toJson()
+      'enterpriseFeatures': enterpriseFeatures?.toJson(),
+      'onfidoTheme': onfidoTheme == null ? null : describeEnum(onfidoTheme)
     };
   }
 }